// workspaces/litellm-orchestrator/plugins/litellm-orchestrator-backend/src/schema/openapi/generated/models/UserAPIKeyAuth.model.ts
/*
 * Copyright 2025 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { Expires } from '../models/Expires.model';

/**
 * Return the row in the db
 * @public
 */
export interface UserAPIKeyAuth {
  token?: string | null;
  key_name?: string | null;
  key_alias?: string | null;
  spend?: number;
  max_budget?: number | null;
  expires?: Expires | null;
  models?: Array<any>;
  aliases?: { [key: string]: any };
  config?: { [key: string]: any };
  user_id?: string | null;
  team_id?: string | null;
  max_parallel_requests?: number | null;
  metadata?: { [key: string]: any };
  tpm_limit?: number | null;
  rpm_limit?: number | null;
  budget_duration?: string | null;
  budget_reset_at?: Date | null;
  allowed_cache_controls?: Array<any> | null;
  allowed_routes?: Array<any> | null;
  permissions?: { [key: string]: any };
  model_spend?: { [key: string]: any };
  model_max_budget?: { [key: string]: any };
  soft_budget_cooldown?: boolean;
  blocked?: boolean | null;
  litellm_budget_table?: { [key: string]: any } | null;
  org_id?: string | null;
  created_at?: Date | null;
  created_by?: string | null;
  updated_at?: Date | null;
  updated_by?: string | null;
  object_permission_id?: string | null;
  object_permission?: { [key: string]: any } | null;
  rotation_count?: number | null;
  auto_rotate?: boolean | null;
  rotation_interval?: string | null;
  last_rotation_at?: Date | null;
  key_rotation_at?: Date | null;
  team_spend?: number | null;
  team_alias?: string | null;
  team_tpm_limit?: number | null;
  team_rpm_limit?: number | null;
  team_max_budget?: number | null;
  team_models?: Array<any>;
  team_blocked?: boolean;
  soft_budget?: number | null;
  team_model_aliases?: { [key: string]: any } | null;
  team_member?: { [key: string]: any } | null;
  team_metadata?: { [key: string]: any } | null;
  team_object_permission_id?: string | null;
  team_member_spend?: number | null;
  team_member_tpm_limit?: number | null;
  team_member_rpm_limit?: number | null;
  end_user_id?: string | null;
  end_user_tpm_limit?: number | null;
  end_user_rpm_limit?: number | null;
  end_user_max_budget?: number | null;
  organization_max_budget?: number | null;
  organization_tpm_limit?: number | null;
  organization_rpm_limit?: number | null;
  organization_metadata?: { [key: string]: any } | null;
  last_refreshed_at?: number | null;
  api_key?: string | null;
  user_role?: string | null;
  allowed_model_region?: UserAPIKeyAuthAllowedModelRegionEnum | null;
  parent_otel_span?: null;
  rpm_limit_per_model?: { [key: string]: number } | null;
  tpm_limit_per_model?: { [key: string]: number } | null;
  user_tpm_limit?: number | null;
  user_rpm_limit?: number | null;
  user_email?: string | null;
  request_route?: string | null;
}

/**
 * @public
 */
export type UserAPIKeyAuthAllowedModelRegionEnum = 'eu' | 'us';